# vulkan: optimize mul_mat for small values of N