… ggml/src/ggml-vulkan/ggml-vulkan.cpp
… ggml/src/ggml-vulkan/vulkan-shaders/mul_mat_vec_q4_k.comp
… ggml/src/ggml-vulkan/vulkan-shaders/mul_mat_vec_q5_k.comp